requirements.txt
asgiref==3.10.0
Django==5.2.7
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
Markdown==3.9
pillow==12.0.0
sqlparse==0.5.3
drf-spectacular==0.29.0
gunicorn==23.0.0
whitenoise==6.11.0
